Add files via upload
You will need the packages for Node.JS as well as puppeteer, you can look up a tutorial for each individual compiler/IDE. I used VS Code for this so it was pretty simple. The code must continuously run on your computer (you can change the time at the bottom for how long you want it to run for in milliseconds) to receive constant updates on your stocks. Happy trading!